auth/requirements.txt
django==4.2.3
gunicorn==21.2.0
djangorestframework==3.14.0
psycopg2-binary==2.9.6